requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
playwright==1.39.0
python-dotenv==1.0.0
openai==1.3.0
pinecone-client==2.2.2
supabase==1.0.3
prometheus-client==0.17.1
requests==2.31.0
python-jose[cryptography]==3.3.0
readability-lxml==0.8.1
psycopg2-binary==2.9.9
pydantic==2.4.2
pytest==7.4.3
ruff==0.1.3
pylint==3.0.2
httpx==0.25.0
python-multipart==0.0.6
starlette==0.27.0
jinja2==3.1.2

# Optional packages
# sentry-sdk==1.32.0 